+++ openapi-codegen/cli/src/main/scala/sttp/tapir/codegen/GenScala.scala
@@ -1,4 +1,4 @@
-package sttp.tapir.cli
+package sttp.tapir.codegen
 
 import cats.data.Validated
 import cats.effect.{IO, ExitCode}
